правки по стилю кода
