+++ js/enums.js
@@ -9,215 +9,4 @@ const AnswerPoints = {
   BONUS: 50
 };
 
-let resultFirst = [
-  {
-    truth: true,
-    time: 35000
-  },
-  {
-    truth: true,
-    time: 32000
-  },
-  {
-    truth: true,
-    time: 7000
-  },
-  {
-    truth: true,
-    time: 25000
-  },
-  {
-    truth: true,
-    time: 12000
-  },
-  {
-    truth: true,
-    time: 17000
-  },
-  {
-    truth: true,
-    time: 22000
-  },
-  {
-    truth: true,
-    time: 12000
-  },
-  {
-    truth: true,
-    time: 25000
-  },
-  {
-    truth: true,
-    time: 11000
-  }
-];
-
-let resultSecond = [
-  {
-    truth: true,
-    time: 8000
-  },
-  {
-    truth: true,
-    time: 8000
-  },
-  {
-    truth: true,
-    time: 7000
-  },
-  {
-    truth: true,
-    time: 8000
-  },
-  {
-    truth: true,
-    time: 8000
-  },
-  {
-    truth: true,
-    time: 8000
-  },
-  {
-    truth: true,
-    time: 8000
-  },
-  {
-    truth: true,
-    time: 8000
-  },
-  {
-    truth: true,
-    time: 8000
-  },
-  {
-    truth: true,
-    time: 8000
-  }
-];
-
-let resultThird = [
-  {
-    truth: true,
-    time: 8000
-  },
-  {
-    truth: true,
-    time: 8000
-  },
-  {
-    truth: true,
-    time: 7000
-  },
-  {
-    truth: true,
-    time: 8000
-  },
-  {
-    truth: true,
-    time: 8000
-  },
-  {
-    truth: true,
-    time: 8000
-  },
-  {
-    truth: true,
-    time: 8000
-  },
-  {
-    truth: true,
-    time: 8000
-  },
-  {
-    truth: true,
-    time: 8000
-  }
-];
-
-let resultFour = [
-  {
-    truth: false,
-    time: 8000
-  },
-  {
-    truth: false,
-    time: 8000
-  },
-  {
-    truth: false,
-    time: 7000
-  },
-  {
-    truth: false,
-    time: 8000
-  },
-  {
-    truth: false,
-    time: 8000
-  },
-  {
-    truth: false,
-    time: 8000
-  },
-  {
-    truth: false,
-    time: 8000
-  },
-  {
-    truth: false,
-    time: 8000
-  },
-  {
-    truth: false,
-    time: 8000
-  },
-  {
-    truth: false,
-    time: 8000
-  }
-];
-
-let resultFifth = [
-  {
-    truth: true,
-    time: 8000
-  },
-  {
-    truth: true,
-    time: 8000
-  },
-  {
-    truth: true,
-    time: 7000
-  },
-  {
-    truth: true,
-    time: 8000
-  },
-  {
-    truth: true,
-    time: 8000
-  },
-  {
-    truth: true,
-    time: 8000
-  },
-  {
-    truth: true,
-    time: 5000
-  },
-  {
-    truth: true,
-    time: 38000
-  },
-  {
-    truth: false,
-    time: 8000
-  },
-  {
-    truth: false,
-    time: 8000
-  }
-];
-
-export {AnswerTime, AnswerPoints, resultFirst, resultSecond, resultThird, resultFour, resultFifth};
+export {AnswerTime, AnswerPoints};
